pyproject.toml
[tool.poetry]
name = "remortalite"
version = "0.1.0"
description = "Presonal site"
authors = ["Rem M. <morrig407@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
pyopenssl = "^24.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
